+++ core/standalone-decorator/index.ts
@@ -1,3 +1,4 @@
 export * from './src/typing';
 export * from './src/util/StandaloneUtil';
 export * from './src/decorator/Runner';
+export * from './src/event/EventHandler';