+++ CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.57.15-alpha.0](https://github.com/eggjs/tegg/compare/v3.57.14...v3.57.15-alpha.0) (2025-06-27)
+
+
+### Bug Fixes
+
+* repo dir ([2e7351b](https://github.com/eggjs/tegg/commit/2e7351bc62161b203f7a98a978d9a901165fb516))
+
+
+### Features
+
+* module plugin ([#325](https://github.com/eggjs/tegg/issues/325)) ([d10bcb6](https://github.com/eggjs/tegg/commit/d10bcb611dbc499ae86558828f91bae3f7266766))
+
+
+
+
+
 ## [3.57.14](https://github.com/eggjs/tegg/compare/v3.57.13...v3.57.14) (2025-06-18)
 
 **Note:** Version bump only for package tegg